refactor: 보다 명확한 네이밍 도입
